+++ module/bitfont/glsl/bitfont.vert
@@ -11,7 +11,7 @@ flat out int vfCharIndex;
 
 
 uniform mat4 matProj;
-uniform float lowResolution;
+uniform int lowResolution;
 
 
 uniform vec2 screenPos;
